Split implementations of clear and clearBlocking because cluster mode needs directed operations. Fix option naming
